Added Mat Dialoge When delete Role

+++ src/app/roles/Components/all-roles/roles.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { RolesService } from '../../Services/roles.service';
 import { Router } from '@angular/router';
 import { Role } from '../../Models/role';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogeComponent } from '../../../Shared/material/dialoge/dialoge.component';
 
 @Component({
   selector: 'app-roles',
@@ -18,7 +20,9 @@ export class RolesComponent implements OnInit  {
   role:Role={} as Role;
 
 
-  constructor(private rolesService:RolesService, private router:Router) { }
+  constructor(private rolesService:RolesService,
+              private router:Router,
+              private dialog: MatDialog) { }
 
 
   ngOnInit(): void {
@@ -38,10 +42,20 @@ export class RolesComponent implements OnInit  {
     this.router.navigate([`roles/add`])
   }
   deleteRole(id:number){
-    // this.openDialog()
-  this.rolesService.deleteRole(id);
 
-  }
+    const dialogRef = this.dialog.open(DialogeComponent, {
+      width: '400px',
+      height:'280px'
+      });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result === 'confirm') {
+
+        this.rolesService.deleteRole(id);
+      }
+      });
+    }
+  
   editRole(id:number){
     this.router.navigate([`roles/edit/${id}`])
